+++ prisma/seed.ts
@@ -5,7 +5,7 @@ const prisma = new PrismaClient();
 async function main() {
   const data: Cats[] = [];
   for (let i = 1; i <= 50; i++) {
-    data.push({ id: i, name: faker.animal.cat(), color: faker.color.human(), weight: faker.random.numeric(), eyes_color: faker.color.human() });
+    data.push({ id: i, name: faker.animal.cat(), color: faker.color.human(), weight: faker.number.int().toString(), eyes_color: faker.color.human() });
   }
 
   const cats = await prisma.cats.createMany({ data, skipDuplicates: true });