+++ src/types/session.d.ts
@@ -1,10 +1,11 @@
 import 'express-session';
+import { Roles } from 'src/guards/is-authorized/roles';
 
 declare module 'express-session' {
   interface Session {
     user?: {
       isLogged: boolean;
-      isAdmin?: boolean;
+      role?: Roles;
 
       // Declare here your session data
     };