+++ src/utils/utils.ts
@@ -1,5 +1,6 @@
 import templateToPDF from './functions/templateToPDF';
+import sendMail from './functions/sendMail';
 import PDFGenerator from './classes/PDFGenerator';
 import * as crypto from './functions/crypto';
 
-export { PDFGenerator, templateToPDF, crypto };
+export { PDFGenerator, templateToPDF, sendMail, crypto };