+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.8.0 - Rework authorization + sendMail utils for mail diversion in dev env
+
+- Rework authorization system
+- Add sendMail utils for mail diversion in dev env
+
 ## v1.7.0 - Production variable in config
 
 - Add production variable in config